Move init code to autoload dir
Autoload functions for starting and stopping vdebug are defined and used
as backends for the :VdebugStart and :VdebugStop commands.

The complete initialisation including loading of the python code is
moved to the autoload file.  This way it will only be loaded when the
debugger is actually started and not on vim startup.